% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RepSeqData}
\alias{RepSeqData}
\title{RepSeqData}
\format{
RepSeqExperiment object and gz-compressed txt/csv tab-delimited :
\describe{
\item{\code{system.file(file.path('extdata/mixcr'), package='RepSeqData')}}{Raw MiXCR aligned data sets}
\item{\code{system.file(file.path('extdata/rtcr'), package='RepSeqData')}}{Raw rTCR aligned data sets}
\item{\code{system.file(file.path('extdata/sampledata.txt'), package='RepSeqData')}}{Sample metadata}
}
}
\source{
\url{https://github.com/ph-pham/RepSeqData}
}
\usage{
RepSeqData
}
\description{
RepSeqExperiment data set built from aligned TCR sequences using MiXCR. MiXCR respecitvely rTCR aligned data are also avaiblable in extdata/mixrc respectively extdata/rtcr
}
\keyword{datasets}
